+++ include/nbl/ext/MitsubaLoader/CMitsubaPipelineMetadata.h
@@ -19,28 +19,30 @@ namespace MitsubaLoader
 //TODO make it inherit from IMitsubaMetadata so that it has global mitsuba metadata ptr
 class CMitsubaPipelineMetadata final : public asset::IRenderpassIndependentPipelineMetadata
 {
-public:
-    CMitsubaPipelineMetadata(core::smart_refctd_ptr<asset::ICPUDescriptorSet>&& _ds0, core::smart_refctd_dynamic_array<ShaderInputSemantic>&& _inputs) :
-        m_ds0(std::move(_ds0)),
-        m_shaderInputs(std::move(_inputs))
-    {
-    }
+    public:
+        CMitsubaPipelineMetadata(core::smart_refctd_ptr<asset::ICPUDescriptorSet>&& _ds0, core::smart_refctd_dynamic_array<ShaderInputSemantic>&& _inputs) :
+            m_ds0(std::move(_ds0)),
+            m_shaderInputs(std::move(_inputs))
+        {
+        }
 
-    core::SRange<const ShaderInputSemantic> getCommonRequiredInputs() const override
-    {
-        return {m_shaderInputs->begin(), m_shaderInputs->end()};
-    }
+        core::SRange<const ShaderInputSemantic> getCommonRequiredInputs() const override
+        {
+            return {m_shaderInputs->begin(), m_shaderInputs->end()};
+        }
 
-    asset::ICPUDescriptorSet* getDescriptorSet() const { return m_ds0.get(); }
+        asset::ICPUDescriptorSet* getDescriptorSet() const { return m_ds0.get(); }
 
-    _NBL_STATIC_INLINE_CONSTEXPR const char* LoaderName = "CMitsubaLoader";
-    const char* getLoaderName() const override { return LoaderName; }
+        _NBL_STATIC_INLINE_CONSTEXPR const char* LoaderName = "CMitsubaLoader";
+        const char* getLoaderName() const override { return LoaderName; }
 
-private:
-    core::smart_refctd_ptr<asset::ICPUDescriptorSet> m_ds0;
-    core::smart_refctd_dynamic_array<ShaderInputSemantic> m_shaderInputs;
+    private:
+        core::smart_refctd_ptr<asset::ICPUDescriptorSet> m_ds0;
+        core::smart_refctd_dynamic_array<ShaderInputSemantic> m_shaderInputs;
 };
 
-}}}
+}
+}
+}
 
 #endif
